% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\name{animals}
\alias{animals}
\title{animals}
\usage{
animals(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO animals
}
\description{
animals
}
\examples{
\dontrun{x <- animals(gdx)}
}
\author{
Abhijeet Mishra
}
