% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biofuelFeedstock.R
\name{biofuelFeedstock}
\alias{biofuelFeedstock}
\title{biofuelFeedstock}
\usage{
biofuelFeedstock(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO biofuelFeedstock
}
\description{
biofuelFeedstock
}
\examples{
\dontrun{x <- biofuelFeedstock(gdx)}
}
\author{
Abhijeet Mishra
}
