% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRelative.R
\name{calcRelative}
\alias{calcRelative}
\title{Calculate relative change for IMPACT indicators}
\usage{
calcRelative(
  df,
  base_year = 2005,
  cols = c("description", "model", "indicator", "region", "unit"),
  type = "same_unit"
)
}
\arguments{
\item{df}{Full IMAPCT result for an indicator}

\item{base_year}{base year on which the relative indicator is calculated}

\item{cols}{columns which are used for relative index calculation}

\item{type}{If relative index is calculate in same unit ('same_unit') or as
an index on base year ('index')}
}
\value{
Recalculate dataframe with relative calculation
}
\description{
Calculate relative change for IMPACT indicators
}
\examples{
\dontrun{x <- calcRelative(df=reportPopulation(gdx))}
}
