% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_description.R
\name{clean_description}
\alias{clean_description}
\title{clean_description}
\usage{
clean_description(df)
}
\arguments{
\item{df}{dataframe with description column}
}
\value{
GLO population
}
\description{
clean_description
}
\examples{
\dontrun{x <- clean_description(df)}
}
\author{
Abhijeet Mishra
}
