% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unit.R
\name{clean_unit}
\alias{clean_unit}
\title{Clean unit}
\usage{
clean_unit(df)
}
\arguments{
\item{df}{dataframe with description column}
}
\value{
Converted units if needed
}
\description{
Clean unit
}
\examples{
\dontrun{x <- clean_unit(df)}
}
\author{
Abhijeet Mishra
}
