% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumerPrices.R
\name{consumerPrices}
\alias{consumerPrices}
\title{Consumer Prices}
\usage{
consumerPrices(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO consumerPrices
}
\description{
Consumer Prices
}
\examples{
\dontrun{x <- consumerPrices(gdx)}
}
\author{
Abhijeet Mishra
}
