% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropArea.R
\name{cropArea}
\alias{cropArea}
\title{cropArea}
\usage{
cropArea(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO cropArea
}
\description{
cropArea
}
\examples{
\dontrun{x <- cropArea(gdx)}
}
\author{
Abhijeet Mishra
}
