% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropAreaAgg.R
\name{cropAreaAgg}
\alias{cropAreaAgg}
\title{cropAreaAgg}
\usage{
cropAreaAgg(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO cropAreaAgg
}
\description{
cropAreaAgg
}
\examples{
\dontrun{x <- cropAreaAgg(gdx)}
}
\author{
Abhijeet Mishra
}
