% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{demand}
\alias{demand}
\title{Demand}
\usage{
demand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO Demand
}
\description{
Demand
}
\examples{
\dontrun{x <- Demand(gdx)}
}
\author{
Abhijeet Mishra
}
