% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticExportPrices.R
\name{DomesticExportPrices}
\alias{DomesticExportPrices}
\title{Domestic Export Prices}
\usage{
DomesticExportPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO DomesticExportPrices
}
\description{
Domestic Export Prices
}
\examples{
\dontrun{x <- DomesticExportPrices(gdx)}
}
\author{
Abhijeet Mishra
}
