% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticImportPrices.R
\name{DomesticImportPrices}
\alias{DomesticImportPrices}
\title{Domestic Import Prices}
\usage{
DomesticImportPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO DomesticImportPrices
}
\description{
Domestic Import Prices
}
\examples{
\dontrun{x <- DomesticImportPrices(gdx)}
}
\author{
Abhijeet Mishra
}
