% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{export}
\usage{
export(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO export
}
\description{
export
}
\examples{
\dontrun{x <- export(gdx)}
}
\author{
Abhijeet Mishra
}
