% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportShareProduction.R
\name{ExportShareProduction}
\alias{ExportShareProduction}
\title{ExportShareProduction}
\usage{
ExportShareProduction(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO ExportShareProduction
}
\description{
ExportShareProduction
}
\examples{
\dontrun{x <- ExportShareProduction(gdx)}
}
\author{
Abhijeet Mishra
}
