% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_locator.R
\name{file_locator}
\alias{file_locator}
\title{File locator}
\usage{
file_locator(source)
}
\arguments{
\item{source}{Path to IMPACT model run}
}
\value{
directory
}
\description{
File locator
}
\examples{
x <- file_locator()

}
\author{
Abhijeet Mishra
}
