% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodAvailability.R
\name{foodAvailability}
\alias{foodAvailability}
\title{FoodAvailability}
\usage{
foodAvailability(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO FoodAvailability
}
\description{
FoodAvailability
}
\examples{
\dontrun{x <- foodAvailability(gdx)}
}
\author{
Abhijeet Mishra
}
