% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDP.R
\name{GDP}
\alias{GDP}
\title{GDP}
\usage{
GDP(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO GDP
}
\description{
GDP
}
\examples{
\dontrun{x <- GDP(gdx)}
}
\author{
Abhijeet Mishra
}
