% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{getReport}
\usage{
getReport(gdx, prep_flag = "Default", export = TRUE)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{prep_flag}{Who processed data}

\item{export}{if RDS file should be written}
}
\value{
Full data output
}
\description{
getReport
}
\examples{
\dontrun{x <- getReport(gdx)}
}
\author{
Abhijeet Mishra
}
