% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdDemand.R
\name{householdDemand}
\alias{householdDemand}
\title{householdDemand}
\usage{
householdDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO householdDemand
}
\description{
householdDemand
}
\examples{
\dontrun{x <- householdDemand(gdx)}
}
\author{
Abhijeet Mishra
}
