% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdIncome.R
\name{householdIncome}
\alias{householdIncome}
\title{householdIncome}
\usage{
householdIncome(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO householdIncome
}
\description{
householdIncome
}
\examples{
\dontrun{x <- householdIncome(gdx)}
}
\author{
Abhijeet Mishra
}
