% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdPopulation.R
\name{householdPopulation}
\alias{householdPopulation}
\title{householdPopulation}
\usage{
householdPopulation(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO householdPopulation
}
\description{
householdPopulation
}
\examples{
\dontrun{x <- householdPopulation(gdx)}
}
\author{
Abhijeet Mishra
}
