% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO import
}
\description{
import
}
\examples{
\dontrun{x <- import(gdx)}
}
\author{
Abhijeet Mishra
}
