% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportShareDemand.R
\name{ImportShareDemand}
\alias{ImportShareDemand}
\title{ImportShareDemand}
\usage{
ImportShareDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO ImportShareDemand
}
\description{
ImportShareDemand
}
\examples{
\dontrun{x <- ImportShareDemand(gdx)}
}
\author{
Abhijeet Mishra
}
