% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediateDemand.R
\name{intermediateDemand}
\alias{intermediateDemand}
\title{Intermediate Demand}
\usage{
intermediateDemand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO other Demand
}
\description{
Intermediate Demand
}
\examples{
\dontrun{x <- intermediateDemand(gdx)}
}
\author{
Abhijeet Mishra
}
