% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelSum.R
\name{levelSum}
\alias{levelSum}
\title{Level based summing}
\usage{
levelSum(
  df,
  dim_name = NULL,
  spatial = "region",
  temporal = "yrs",
  na.rm = TRUE,
  bind = TRUE
)
}
\arguments{
\item{df}{A dataframe which comes out of aggregateIMPACT function call}

\item{dim_name}{The column to be summed over. Usually "long_name"}

\item{spatial}{Which column contains spatial info. Defaults to "region"}

\item{temporal}{Which column contains temporal info. Defaults to "yrs"}

\item{na.rm}{If NAs should be ignored during sum. Defaults to TRUE}

\item{bind}{If new aggregation should be binded to the original dataframe
passed on to "df" argument}
}
\value{
Aggregated dataframe
}
\description{
Level based summing
}
\examples{
\dontrun{x <- levelSum(df)}

}
