% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsfDemand.R
\name{lsfDemand}
\alias{lsfDemand}
\title{lsfDemand}
\usage{
lsfDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO lsfDemand
}
\description{
lsfDemand
}
\examples{
\dontrun{x <- lsfDemand(gdx)}
}
\author{
Abhijeet Mishra
}
