% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malnourished.R
\name{malnourished}
\alias{malnourished}
\title{Malnourished}
\usage{
malnourished(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO malnourished
}
\description{
Malnourished
}
\examples{
\dontrun{x <- malnourished(gdx)}
}
\author{
Abhijeet Mishra
}
