% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetPrices.R
\name{NetPrices}
\alias{NetPrices}
\title{Net Prices}
\usage{
NetPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO NetPrices
}
\description{
Net Prices
}
\examples{
\dontrun{x <- NetPrices(gdx)}
}
\author{
Abhijeet Mishra
}
