% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTrade.R
\name{netTrade}
\alias{netTrade}
\title{netTrade}
\usage{
netTrade(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO netExport
}
\description{
netTrade
}
\examples{
\dontrun{x <- netExport(gdx)}
}
\author{
Abhijeet Mishra
}
