% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetTradeShareDemand.R
\name{NetTradeShareDemand}
\alias{NetTradeShareDemand}
\title{NetTradeShareDemand}
\usage{
NetTradeShareDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO NetTradeShareDemand
}
\description{
NetTradeShareDemand
}
\examples{
\dontrun{x <- NetTradeShareDemand(gdx)}
}
\author{
Abhijeet Mishra
}
