% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetTradeShareProduction.R
\name{NetTradeShareProduction}
\alias{NetTradeShareProduction}
\title{NetTradeShareProduction}
\usage{
NetTradeShareProduction(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO NetTradeShareProduction
}
\description{
NetTradeShareProduction
}
\examples{
\dontrun{x <- NetTradeShareProduction(gdx)}
}
\author{
Abhijeet Mishra
}
