% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherDemand.R
\name{otherDemand}
\alias{otherDemand}
\title{Other Demand}
\usage{
otherDemand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO other Demand
}
\description{
Other Demand
}
\examples{
\dontrun{x <- otherDemand(gdx)}
}
\author{
Abhijeet Mishra
}
