% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerCapKCal.R
\name{PerCapKCal}
\alias{PerCapKCal}
\title{KCal per Capita}
\usage{
PerCapKCal(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
Per capita KCal availability
}
\description{
KCal per Capita
}
\examples{
\dontrun{x <- PerCapKCal(gdx)}
}
\author{
Abhijeet Mishra
}
