% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerCapKCalCommodity.R
\name{PerCapKCalCommodity}
\alias{PerCapKCalCommodity}
\title{KCal per Capita per Commodity}
\usage{
PerCapKCalCommodity(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
Per capita KCal availability
}
\description{
KCal per Capita per Commodity
}
\examples{
\dontrun{x <- PerCapKCalCommodity(gdx)}
}
\author{
Abhijeet Mishra
}
