% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Population}
\usage{
population(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
Population
}
\examples{
population_glo <- population(gdx)
}
\author{
Abhijeet Mishra
}
