% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAggregator.R
\name{postAggregator}
\alias{postAggregator}
\title{Post Aggregator}
\usage{
postAggregator(df, gdx, col = "indicator", type = "crop_area")
}
\arguments{
\item{df}{Dataframe which is received from clean_description}

\item{gdx}{A gdx file to check for the standard columns}

\item{col}{The "indicator" column}

\item{type}{which kind of post aggregation is needed}
}
\value{
Dataframe with correct aggregation to subscales (upper level aggregation)
}
\description{
Post Aggregator
}
\examples{
\dontrun{x <- postAggregator(df)}
}
