% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{production}
\alias{production}
\title{Production}
\usage{
production(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO Production
}
\description{
Production
}
\examples{
\dontrun{x <- Production(gdx)}
}
\author{
Abhijeet Mishra
}
