% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAnimals.R
\name{reportAnimals}
\alias{reportAnimals}
\title{reportAnimals}
\usage{
reportAnimals(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO Animals
}
\description{
reportAnimals
}
\examples{
\dontrun{x <- reportAnimals(gdx)}
}
\author{
Abhijeet Mishra
}
