% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBiofuelFeedstock.R
\name{reportBiofuelFeedstock}
\alias{reportBiofuelFeedstock}
\title{reportBiofuelFeedstock}
\usage{
reportBiofuelFeedstock(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO BiofuelFeedstock
}
\description{
reportBiofuelFeedstock
}
\examples{
\dontrun{x <- reportBiofuelFeedstock(gdx)}
}
\author{
Abhijeet Mishra
}
