% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportConsumerPrices.R
\name{reportConsumerPrices}
\alias{reportConsumerPrices}
\title{reportConsumerPrices}
\usage{
reportConsumerPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportConsumerPrices
}
\examples{
\dontrun{x <- reportConsumerPrices(gdx)}
}
\author{
Abhijeet Mishra
}
