% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropArea.R
\name{reportCropArea}
\alias{reportCropArea}
\title{reportCropArea}
\usage{
reportCropArea(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO CropArea
}
\description{
reportCropArea
}
\examples{
\dontrun{x <- reportCropArea(gdx)}
}
\author{
Abhijeet Mishra
}
