% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemand.R
\name{reportDemand}
\alias{reportDemand}
\title{reportDemand}
\usage{
reportDemand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportDemand
}
\examples{
\dontrun{x <- reportDemand(gdx)}
}
\author{
Abhijeet Mishra
}
