% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDomesticExportPrices.R
\name{reportDomesticExportPrices}
\alias{reportDomesticExportPrices}
\title{reportDomesticExportPrices}
\usage{
reportDomesticExportPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportDomesticExportPrices
}
\examples{
\dontrun{x <- reportDomesticExportPrices(gdx)}
}
\author{
Abhijeet Mishra
}
