% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDomesticImportPrices.R
\name{reportDomesticImportPrices}
\alias{reportDomesticImportPrices}
\title{reportDomesticImportPrices}
\usage{
reportDomesticImportPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportDomesticImportPrices
}
\examples{
\dontrun{x <- reportDomesticImportPrices(gdx)}
}
\author{
Abhijeet Mishra
}
