% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExport.R
\name{reportExport}
\alias{reportExport}
\title{reportExport}
\usage{
reportExport(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO export
}
\description{
reportExport
}
\examples{
\dontrun{x <- reportexport(gdx)}
}
\author{
Abhijeet Mishra
}
