% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExportShareProduction.R
\name{reportExportShareProduction}
\alias{reportExportShareProduction}
\title{reportExportShareProduction}
\usage{
reportExportShareProduction(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportExportShareProduction
}
\examples{
\dontrun{x <- reportExportShareProduction(gdx)}
}
\author{
Abhijeet Mishra
}
