% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFoodAvailability.R
\name{reportFoodAvailability}
\alias{reportFoodAvailability}
\title{reportFoodAvailability}
\usage{
reportFoodAvailability(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO FoodAvailability
}
\description{
reportFoodAvailability
}
\examples{
\dontrun{x <- reportFoodAvailability(gdx)}
}
\author{
Abhijeet Mishra
}
