% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGDP.R
\name{reportGDP}
\alias{reportGDP}
\title{reportGDP}
\usage{
reportGDP(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportGDP
}
\examples{
\dontrun{x <- reportGDP(gdx)}
}
\author{
Abhijeet Mishra
}
