% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHouseholdDemand.R
\name{reportHouseholdDemand}
\alias{reportHouseholdDemand}
\title{reportHouseholdDemand}
\usage{
reportHouseholdDemand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportHouseholdDemand
}
\examples{
\dontrun{x <- reportHouseholdDemand(gdx)}
}
\author{
Abhijeet Mishra
}
