% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHouseholdIncome.R
\name{reportHouseholdIncome}
\alias{reportHouseholdIncome}
\title{reportHouseholdIncome}
\usage{
reportHouseholdIncome(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportHouseholdIncome
}
\examples{
\dontrun{x <- reportHouseholdIncome(gdx)}
}
\author{
Abhijeet Mishra
}
