% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHouseholdPopulation.R
\name{reportHouseholdPopulation}
\alias{reportHouseholdPopulation}
\title{reportHouseholdPopulation}
\usage{
reportHouseholdPopulation(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportHouseholdPopulation
}
\examples{
\dontrun{x <- reportHouseholdPopulation(gdx)}
}
\author{
Abhijeet Mishra
}
