% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHungerRisk.R
\name{reportHungerRisk}
\alias{reportHungerRisk}
\title{reportHungerRisk}
\usage{
reportHungerRisk(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportHungerRisk
}
\examples{
\dontrun{x <- reportHungerRisk(gdx)}
}
\author{
Abhijeet Mishra
}
