% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportImport.R
\name{reportImport}
\alias{reportImport}
\title{reportImport}
\usage{
reportImport(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO export
}
\description{
reportImport
}
\examples{
\dontrun{x <- reportImport(gdx)}
}
\author{
Abhijeet Mishra
}
