% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportImportShareDemand.R
\name{reportImportShareDemand}
\alias{reportImportShareDemand}
\title{reportImportShareDemand}
\usage{
reportImportShareDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportImportShareDemand
}
\examples{
\dontrun{x <- reportImportShareDemand(gdx)}
}
\author{
Abhijeet Mishra
}
