% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLSFDemand.R
\name{reportLSFDemand}
\alias{reportLSFDemand}
\title{reportLSFDemand}
\usage{
reportLSFDemand(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportLSFDemand
}
\examples{
\dontrun{x <- reportLSFDemand(gdx)}
}
\author{
Abhijeet Mishra
}
