% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMalnourished.R
\name{reportMalnourished}
\alias{reportMalnourished}
\title{reportMalnourished}
\usage{
reportMalnourished(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO Malnourished
}
\description{
reportMalnourished
}
\examples{
\dontrun{x <- reportMalnourished(gdx)}
}
\author{
Abhijeet Mishra
}
