% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNetPrices.R
\name{reportNetPrices}
\alias{reportNetPrices}
\title{reportNetPrices}
\usage{
reportNetPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
Solution Net Prices
}
\description{
reportNetPrices
}
\examples{
\dontrun{x <- reportNetPrices(gdx)}
}
\author{
Abhijeet Mishra
}
