% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNetTrade.R
\name{reportNetTrade}
\alias{reportNetTrade}
\title{reportNetTrade}
\usage{
reportNetTrade(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO NetExport
}
\description{
reportNetTrade
}
\examples{
\dontrun{x <- reportNetExport(gdx)}
}
\author{
Abhijeet Mishra
}
