% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNetTradeShareDemand.R
\name{reportNetTradeShareDemand}
\alias{reportNetTradeShareDemand}
\title{reportNetTradeShareDemand}
\usage{
reportNetTradeShareDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportNetTradeShareDemand
}
\examples{
\dontrun{x <- reportNetTradeShareDemand(gdx)}
}
\author{
Abhijeet Mishra
}
