% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNetTradeShareProduction.R
\name{reportNetTradeShareProduction}
\alias{reportNetTradeShareProduction}
\title{reportNetTradeShareProduction}
\usage{
reportNetTradeShareProduction(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportNetTradeShareProduction
}
\examples{
\dontrun{x <- reportNetTradeShareProduction(gdx)}
}
\author{
Abhijeet Mishra
}
