% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerCapGDP.R
\name{reportPerCapGDP}
\alias{reportPerCapGDP}
\title{reportPerCapGDP}
\usage{
reportPerCapGDP(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportPerCapGDP
}
\examples{
\dontrun{x <- reportPerCapGDP(gdx)}
}
\author{
Abhijeet Mishra
}
