% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerCapKCal.R
\name{reportPerCapKCal}
\alias{reportPerCapKCal}
\title{reportPerCapKCal}
\usage{
reportPerCapKCal(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO FoodAvailability
}
\description{
reportPerCapKCal
}
\examples{
\dontrun{x <- reportPerCapKCal(gdx)}
}
\author{
Abhijeet Mishra
}
