% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerCapKCalCommodity.R
\name{reportPerCapKCalCommodity}
\alias{reportPerCapKCalCommodity}
\title{reportPerCapKCalCommodity}
\usage{
reportPerCapKCalCommodity(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO FoodAvailability
}
\description{
reportPerCapKCalCommodity
}
\examples{
\dontrun{x <- reportPerCapKCalCommodity(gdx)}
}
\author{
Abhijeet Mishra
}
