% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPopulation.R
\name{reportPopulation}
\alias{reportPopulation}
\title{reportPopulation}
\usage{
reportPopulation(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO population
}
\description{
reportPopulation
}
\examples{
\dontrun{x <- reportPopulation(gdx)}
}
\author{
Abhijeet Mishra
}
