% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSupply.R
\name{reportSupply}
\alias{reportSupply}
\title{reportSupply}
\usage{
reportSupply(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportSupply
}
\examples{
\dontrun{x <- reportSupply(gdx)}
}
\author{
Abhijeet Mishra
}
