% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWeightedWorldPrices.R
\name{reportWeightedWorldPrices}
\alias{reportWeightedWorldPrices}
\title{reportWeightedWorldPrices}
\usage{
reportWeightedWorldPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportWeightedWorldPrices
}
\examples{
\dontrun{x <- reportWeightedWorldPrices(gdx)}
}
\author{
Abhijeet Mishra
}
