% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYields.R
\name{reportYields}
\alias{reportYields}
\title{reportYields}
\usage{
reportYields(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO Yield
}
\description{
reportYields
}
\examples{
\dontrun{x <- reportYield(gdx)}
}
\author{
Abhijeet Mishra
}
