% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleWorldPrices.R
\name{singleWorldPrices}
\alias{singleWorldPrices}
\title{Single World Prices}
\usage{
singleWorldPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO singleWorldPrices
}
\description{
Single World Prices
}
\examples{
\dontrun{x <- singleWorldPrices(gdx)}
}
\author{
Abhijeet Mishra
}
