% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedWorldPrices.R
\name{weightedWorldPrices}
\alias{weightedWorldPrices}
\title{World Prices}
\usage{
weightedWorldPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO weightedWorldPrices
}
\description{
World Prices
}
\examples{
\dontrun{x <- weightedWorldPrices(gdx)}
}
\author{
Abhijeet Mishra
}
