% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yields.R
\name{yields}
\alias{yields}
\title{Yields}
\usage{
yields(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO Yields
}
\description{
Yields
}
\examples{
\dontrun{x <- Yields(gdx)}
}
\author{
Abhijeet Mishra
}
