% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{getReport}
\usage{
getReport(
  gdx,
  additional_indicators = FALSE,
  prep_flag = NULL,
  export = TRUE,
  base_year = NULL,
  sp_mapping = "Standard-IMPACT_dis1",
  relative_calc = FALSE
)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{additional_indicators}{If non-core indicators should be read.}

\item{prep_flag}{Who processed data.
Defaults to user name from the computer where this script is run.}

\item{export}{if RDS file should be written. Defaults to TRUE}

\item{base_year}{Base year on which relative index can be calculated.
Default 2005.}

\item{sp_mapping}{Which mapping to use. See mapping file in DOORMAT package.}

\item{relative_calc}{If relative calculations should be done}
}
\value{
Full data output
}
\description{
getReport
}
\examples{
\dontrun{x <- getReport(gdx)}
}
\author{
Abhijeet Mishra
}
