% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hungerRisk.R
\name{hungerRisk}
\alias{hungerRisk}
\title{hungerRisk}
\usage{
hungerRisk(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO hungerRisk
}
\description{
hungerRisk
}
\examples{
\dontrun{x <- hungerRisk(gdx)}
}
\author{
Abhijeet Mishra
}
