% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cleaner.R
\name{name_cleaner}
\alias{name_cleaner}
\title{Name Cleaner}
\usage{
name_cleaner(df)
}
\arguments{
\item{df}{placeholder dataframe for cleanup}
}
\value{
cleaned dataframe with right names
}
\description{
Name Cleaner
}
\examples{
\dontrun{x <- name_cleaner(df)}
}
