% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percapGDP.R
\name{percapGDP}
\alias{percapGDP}
\title{percapGDP}
\usage{
percapGDP(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO percapGDP
}
\description{
percapGDP
}
\examples{
\dontrun{x <- percapGDP(gdx)}
}
\author{
Abhijeet Mishra
}
