% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producerPrices.R
\name{producerPrices}
\alias{producerPrices}
\title{Producer Prices}
\usage{
producerPrices(gdx)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
GLO producerPrices
}
\description{
Producer Prices
}
\examples{
\dontrun{x <- producerPrices(gdx)}
}
\author{
Abhijeet Mishra
}
