% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIntermediateDemand.R
\name{reportIntermediateDemand}
\alias{reportIntermediateDemand}
\title{reportIntermediateDemand}
\usage{
reportIntermediateDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportIntermediateDemand
}
\examples{
\dontrun{x <- reportIntermediateDemand(gdx)}
}
\author{
Abhijeet Mishra
}
