% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOtherDemand.R
\name{reportOtherDemand}
\alias{reportOtherDemand}
\title{reportOtherDemand}
\usage{
reportOtherDemand(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportOtherDemand
}
\examples{
\dontrun{x <- reportOtherDemand(gdx)}
}
\author{
Abhijeet Mishra
}
