% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProducerPrices.R
\name{reportProducerPrices}
\alias{reportProducerPrices}
\title{reportProducerPrices}
\usage{
reportProducerPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportProducerPrices
}
\examples{
\dontrun{x <- reportproducerPrices(gdx)}
}
\author{
Abhijeet Mishra
}
