% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProduction.R
\name{reportProduction}
\alias{reportProduction}
\title{reportProduction}
\usage{
reportProduction(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO population
}
\description{
reportProduction
}
\examples{
\dontrun{x <- reportProduction(gdx)}
}
\author{
Abhijeet Mishra
}
