% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSingleWorldPrices.R
\name{reportSingleWorldPrices}
\alias{reportSingleWorldPrices}
\title{reportSingleWorldPrices}
\usage{
reportSingleWorldPrices(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO CropArea
}
\description{
reportSingleWorldPrices
}
\examples{
\dontrun{x <- reportSingleWorldPrices(gdx)}
}
\author{
Abhijeet Mishra
}
