% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{supply}
\alias{supply}
\title{Supply}
\usage{
supply(gdx, ...)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{...}{Arguments to aggregateIMPACT call. See ?aggregateIMPACT}
}
\value{
GLO supply
}
\description{
Supply
}
\examples{
\dontrun{x <- supply(gdx)}
}
\author{
Abhijeet Mishra
}
